Fixed autocommit calls for mysql-connector-python
The MySQLdb and mysql-connector-python clients use different methods for getting and setting the autocommit mode. This code checks the `conn` param passed into get/set_autocommit() to see if it's a MySQLdb instance (containing the get_autocommit method) or a mysql-connector-python instance (missing get_autocommit) and makes the appropriate autocommit calls for the client.